+++ openthread/src/radio/spinel.rs
@@ -61,7 +61,7 @@ use crate::sys::OT_RADIO_FRAME_MAX_SIZE;
 /// - Over **SPI** the frames are *not* HDLC-framed: each SPI transaction carries
 ///   a 5-byte header whose length field delimits the payload, so the spinel
 ///   frame is carried raw. That is what [`SpiSpinelTransport`] does — and it also
-///   handles SPI's lack of a slave-initiated channel via an interrupt line.
+///   handles SPI's lack of a peripheral-initiated channel via an interrupt line.
 ///
 /// So `SpinelRadio` builds and parses only *raw* spinel frames and is agnostic to
 /// framing; each transport frames however its wire requires.
@@ -161,159 +161,18 @@ const RESPONSE_TIMEOUT: Duration = Duration::from_millis(1000);
 const MAX_SPINEL_FRAME: usize = OT_RADIO_FRAME_MAX_SIZE as usize + 128;
 
 // ---------------------------------------------------------------------------
-// HDLC framing (RFC 1662 byte-stuffing + CRC-16/X.25 FCS).
+// Transports: putting a raw spinel frame on a concrete wire.
 //
-// Used only by `UartSpinelTransport` (a UART is a raw byte stream with no
-// framing of its own). SPI does not use HDLC — see `SpiSpinelTransport`.
+// `SpinelRadio` is transport-agnostic; each of these implements
+// `SpinelTransport` for a specific bus and owns that bus's framing (HDLC for a
+// UART, the 5-byte header protocol for SPI).
 // ---------------------------------------------------------------------------
 
-const HDLC_FLAG: u8 = 0x7e;
-const HDLC_ESCAPE: u8 = 0x7d;
-const HDLC_XOR: u8 = 0x20;
-const HDLC_FCS_INIT: u16 = 0xffff;
-const HDLC_FCS_GOOD: u16 = 0xf0b8;
+pub mod spi;
+pub mod uart;
 
-/// CRC-16/X.25 (reflected 0x1021) FCS, one byte at a time.
-fn hdlc_fcs_update(fcs: u16, byte: u8) -> u16 {
-    let mut fcs = fcs;
-    let mut b = byte;
-    for _ in 0..8 {
-        if ((fcs ^ (b as u16)) & 0x0001) != 0 {
-            fcs = (fcs >> 1) ^ 0x8408;
-        } else {
-            fcs >>= 1;
-        }
-        b >>= 1;
-    }
-    fcs
-}
-
-fn hdlc_needs_escape(byte: u8) -> bool {
-    matches!(byte, HDLC_FLAG | HDLC_ESCAPE) || byte == 0x11 || byte == 0x13
-}
-
-/// HDLC-encode `payload` into `out`, returning the encoded length (or `None` if
-/// `out` is too small). Wraps with flags, byte-stuffs, and appends the FCS.
-fn hdlc_encode(payload: &[u8], out: &mut [u8]) -> Option<usize> {
-    let mut n = 0;
-
-    let push = |out: &mut [u8], n: &mut usize, byte: u8| -> bool {
-        if hdlc_needs_escape(byte) {
-            if *n + 2 > out.len() {
-                return false;
-            }
-            out[*n] = HDLC_ESCAPE;
-            out[*n + 1] = byte ^ HDLC_XOR;
-            *n += 2;
-        } else {
-            if *n + 1 > out.len() {
-                return false;
-            }
-            out[*n] = byte;
-            *n += 1;
-        }
-        true
-    };
-
-    if n >= out.len() {
-        return None;
-    }
-    out[n] = HDLC_FLAG;
-    n += 1;
-
-    let mut fcs = HDLC_FCS_INIT;
-    for &b in payload {
-        fcs = hdlc_fcs_update(fcs, b);
-        if !push(out, &mut n, b) {
-            return None;
-        }
-    }
-
-    // Append the (complemented) FCS, little-endian, byte-stuffed.
-    fcs ^= 0xffff;
-    if !push(out, &mut n, (fcs & 0xff) as u8) {
-        return None;
-    }
-    if !push(out, &mut n, (fcs >> 8) as u8) {
-        return None;
-    }
-
-    if n >= out.len() {
-        return None;
-    }
-    out[n] = HDLC_FLAG;
-    n += 1;
-
-    Some(n)
-}
-
-/// Incremental HDLC decoder: fed bytes from the transport, yields complete,
-/// FCS-checked spinel frames.
-struct HdlcDecoder {
-    buf: [u8; MAX_SPINEL_FRAME],
-    len: usize,
-    fcs: u16,
-    escaped: bool,
-    in_frame: bool,
-}
-
-impl HdlcDecoder {
-    const fn new() -> Self {
-        Self {
-            buf: [0; MAX_SPINEL_FRAME],
-            len: 0,
-            fcs: HDLC_FCS_INIT,
-            escaped: false,
-            in_frame: false,
-        }
-    }
-
-    fn reset(&mut self) {
-        self.len = 0;
-        self.fcs = HDLC_FCS_INIT;
-        self.escaped = false;
-        self.in_frame = true;
-    }
-
-    /// Push one received byte. Returns `Some(len)` when a complete, valid frame
-    /// is available in `self.buf[..len - 2]` (payload, FCS stripped).
-    fn push(&mut self, byte: u8) -> Option<usize> {
-        match byte {
-            HDLC_FLAG => {
-                let complete = self.in_frame && self.len >= 2 && self.fcs == HDLC_FCS_GOOD;
-
-                let payload_len = self.len.saturating_sub(2);
-                self.reset();
-
-                complete.then_some(payload_len)
-            }
-            HDLC_ESCAPE => {
-                self.escaped = true;
-                None
-            }
-            _ => {
-                if !self.in_frame {
-                    return None;
-                }
-                let b = if self.escaped {
-                    self.escaped = false;
-                    byte ^ HDLC_XOR
-                } else {
-                    byte
-                };
-                if self.len < self.buf.len() {
-                    self.buf[self.len] = b;
-                    self.len += 1;
-                    self.fcs = hdlc_fcs_update(self.fcs, b);
-                } else {
-                    // Overflow — drop the frame.
-                    self.in_frame = false;
-                }
-                None
-            }
-        }
-    }
-}
+pub use spi::{SpiSpinelTransport, SpiTransportError};
+pub use uart::{UartSpinelTransport, UartTransportError};
 
 // ---------------------------------------------------------------------------
 // Spinel frame build/parse.
@@ -907,339 +766,4 @@ where
             // Other frames (matched responses to a concurrent op, status) — ignore.
         }
     }
-}
-
-// ===========================================================================
-// UartSpinelTransport: spinel frames over a UART (HDLC framing).
-// ===========================================================================
-
-/// A [`SpinelTransport`] over a UART (or any `embedded-io-async` byte stream).
-///
-/// A UART is a raw, unframed byte stream, so this transport HDLC-frames each
-/// spinel frame on the way out (byte-stuffing + FCS) and runs an incremental
-/// HDLC decoder on the way in — exactly OpenThread's POSIX `hdlc_interface`.
-///
-/// Wrap any full-duplex byte stream implementing [`embedded_io_async::Read`] +
-/// [`embedded_io_async::Write`] (e.g. an embassy UART):
-///
-/// ```ignore
-/// let radio = SpinelRadio::new(UartSpinelTransport::new(uart));
-/// ```
-pub struct UartSpinelTransport<U> {
-    uart: U,
-    /// HDLC-encode scratch (worst case: every byte escaped, plus flags + FCS).
-    tx_hdlc: [u8; MAX_SPINEL_FRAME * 2 + 4],
-    /// Incremental decoder for the inbound byte stream.
-    decoder: HdlcDecoder,
-    /// Read scratch pulled from the UART in chunks.
-    rx_chunk: [u8; 128],
-    /// Bytes buffered in `rx_chunk` not yet fed to the decoder, `[pos, fill)`.
-    rx_pos: usize,
-    rx_fill: usize,
-}
-
-impl<U> UartSpinelTransport<U> {
-    /// Create a UART spinel transport over `uart`.
-    pub const fn new(uart: U) -> Self {
-        Self {
-            uart,
-            tx_hdlc: [0; MAX_SPINEL_FRAME * 2 + 4],
-            decoder: HdlcDecoder::new(),
-            rx_chunk: [0; 128],
-            rx_pos: 0,
-            rx_fill: 0,
-        }
-    }
-}
-
-impl<U> SpinelTransport for UartSpinelTransport<U>
-where
-    U: embedded_io_async::Read + embedded_io_async::Write,
-{
-    type Error = UartTransportError<U::Error>;
-
-    async fn send(&mut self, frame: &[u8]) -> Result<(), Self::Error> {
-        let n = hdlc_encode(frame, &mut self.tx_hdlc).ok_or(UartTransportError::FrameTooLarge)?;
-        self.uart
-            .write_all(&self.tx_hdlc[..n])
-            .await
-            .map_err(UartTransportError::Io)
-    }
-
-    async fn recv(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        loop {
-            // Drain whatever is already buffered, feeding the HDLC decoder.
-            while self.rx_pos < self.rx_fill {
-                let byte = self.rx_chunk[self.rx_pos];
-                self.rx_pos += 1;
-
-                if let Some(len) = self.decoder.push(byte) {
-                    let len = len.min(buf.len());
-                    buf[..len].copy_from_slice(&self.decoder.buf[..len]);
-                    return Ok(len);
-                }
-            }
-
-            // Refill from the UART.
-            let n = self
-                .uart
-                .read(&mut self.rx_chunk)
-                .await
-                .map_err(UartTransportError::Io)?;
-            if n == 0 {
-                return Err(UartTransportError::Eof);
-            }
-            self.rx_pos = 0;
-            self.rx_fill = n;
-        }
-    }
-}
-
-/// Error type for [`UartSpinelTransport`].
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum UartTransportError<E> {
-    /// The underlying UART returned an error.
-    Io(E),
-    /// The UART reached end-of-stream (a `read` returned 0 bytes).
-    Eof,
-    /// A spinel frame exceeded the HDLC encode buffer.
-    FrameTooLarge,
-}
-
-// ===========================================================================
-// SpiSpinelTransport: spinel frames over SPI (5-byte SPI header protocol).
-// ===========================================================================
-
-/// SPI framing header size (flag byte + accept-len + data-len, all before the
-/// payload). See [`SpiSpinelTransport`] for the layout.
-const SPI_HEADER_SIZE: usize = 5;
-
-/// Flag-byte constants (from OpenThread's `spi_frame.hpp`).
-const SPI_FLAG_RESET: u8 = 1 << 7;
-const SPI_FLAG_PATTERN: u8 = 0x02;
-const SPI_FLAG_PATTERN_MASK: u8 = 0x03;
-
-/// A minimum SPI transfer size, so a small frame the RCP wants to send us can be
-/// picked up in a single transaction even before we know its length.
-const SPI_SMALL_PACKET_SIZE: usize = 32;
-
-/// A [`SpinelTransport`] over SPI, using OpenThread's SPI framing protocol
-/// (`spi_frame.hpp` / the POSIX `spi_interface`), against a stock `ot-rcp` built
-/// with the SPI HDLC-less spinel interface.
-///
-/// # Why SPI needs more than a byte pipe
-///
-/// SPI is master-driven and full-duplex-per-transfer: the RCP (a SPI *slave*)
-/// cannot start a transfer, and every clocked byte shifts one out *and* one in.
-/// So two extra mechanisms are layered on the bus:
-///
-/// - **An interrupt line** (`int`, active-low): the RCP asserts it to say "I have
-///   a frame for you, please clock a transfer". This is the slave-initiated
-///   channel raw SPI lacks. This transport [`Wait`](embedded_hal_async::digital::Wait)s
-///   on it before reading.
-/// - **A 5-byte SPI frame header** on every transfer:
-///   `flag(1) | accept_len(2, LE) | data_len(2, LE)`. `accept_len` is how many
-///   payload bytes *this* side can receive; `data_len` is how many it is sending.
-///   Because the transfer is full-duplex, one transaction simultaneously pushes
-///   our frame and pulls the RCP's, each bounded by the other's advertised
-///   `accept_len`. The flag byte carries a reset bit and a fixed pattern
-///   (`0x02`) that distinguishes a real header from an idle `0x00`/`0xFF` bus.
-///
-/// The payload inside the SPI frame is the **raw spinel frame** — SPI does its
-/// own framing via `data_len`, so (unlike UART) there is no HDLC.
-///
-/// ```ignore
-/// // `spi: embedded_hal_async::spi::SpiDevice`, `int: embedded_hal_async::digital::Wait`
-/// let radio = SpinelRadio::new(SpiSpinelTransport::new(spi, int));
-/// ```
-pub struct SpiSpinelTransport<S, I> {
-    spi: S,
-    int: I,
-    /// Full SPI transfer buffers (header + payload), reused every transaction.
-    tx_buf: [u8; SPI_HEADER_SIZE + MAX_SPINEL_FRAME],
-    rx_buf: [u8; SPI_HEADER_SIZE + MAX_SPINEL_FRAME],
-    /// A single queued outbound spinel frame waiting to be accepted by the RCP.
-    tx_pending: Option<usize>,
-    /// A single received spinel frame not yet handed to `recv`, `[.. len]` of
-    /// `rx_frame`.
-    rx_frame: [u8; MAX_SPINEL_FRAME],
-    rx_ready: Option<usize>,
-    /// The RCP's last-advertised pending data length (learned from a prior header
-    /// so the next transfer is sized to fit it).
-    slave_data_len: usize,
-    /// Whether we still owe the RCP the reset flag on our next header (set on the
-    /// very first transfer so the RCP knows the host came up fresh).
-    send_reset: bool,
-}
-
-impl<S, I> SpiSpinelTransport<S, I> {
-    /// Create an SPI spinel transport over SPI device `spi` and interrupt line
-    /// `int` (active-low: asserted low when the RCP has data to send).
-    pub const fn new(spi: S, int: I) -> Self {
-        Self {
-            spi,
-            int,
-            tx_buf: [0; SPI_HEADER_SIZE + MAX_SPINEL_FRAME],
-            rx_buf: [0; SPI_HEADER_SIZE + MAX_SPINEL_FRAME],
-            tx_pending: None,
-            rx_frame: [0; MAX_SPINEL_FRAME],
-            rx_ready: None,
-            slave_data_len: 0,
-            send_reset: true,
-        }
-    }
-}
-
-impl<S, I> SpiSpinelTransport<S, I>
-where
-    S: embedded_hal_async::spi::SpiDevice,
-    I: embedded_hal_async::digital::Wait,
-{
-    /// Perform one SPI transaction, honoring flow control in both directions:
-    /// advertise how much we can receive, send our queued frame (if any), and
-    /// pick up the RCP's frame (if it fits). Returns `Ok(true)` if a received
-    /// frame was buffered into `rx_frame`.
-    async fn push_pull(&mut self) -> Result<bool, SpiTransportError<S::Error, I::Error>> {
-        // ---- Build our TX header + payload. ----
-        let tx_payload_len = self.tx_pending.unwrap_or(0);
-
-        // How much we're willing to receive this transfer: enough for the RCP's
-        // announced pending frame, or at least a small-packet floor.
-        let accept_len = if self.slave_data_len != 0 {
-            self.slave_data_len
-        } else {
-            SPI_SMALL_PACKET_SIZE
-        }
-        .max(tx_payload_len);
-
-        // Flag byte: fixed pattern, plus the reset bit on the first transfer.
-        self.tx_buf[0] = SPI_FLAG_PATTERN | if self.send_reset { SPI_FLAG_RESET } else { 0 };
-        self.tx_buf[1..3].copy_from_slice(&(accept_len as u16).to_le_bytes());
-        self.tx_buf[3..5].copy_from_slice(&(tx_payload_len as u16).to_le_bytes());
-
-        // The payload was staged into `tx_buf[SPI_HEADER_SIZE..]` by `send`.
-        let transfer_len = SPI_HEADER_SIZE + accept_len.max(tx_payload_len);
-        if transfer_len > self.tx_buf.len() {
-            return Err(SpiTransportError::FrameTooLarge);
-        }
-
-        // Zero the unused RX region so a short/no-op slave reply reads as zeros.
-        for b in self.rx_buf[..transfer_len].iter_mut() {
-            *b = 0;
-        }
-
-        // ---- Clock the full-duplex transfer. ----
-        self.spi
-            .transfer(
-                &mut self.rx_buf[..transfer_len],
-                &self.tx_buf[..transfer_len],
-            )
-            .await
-            .map_err(SpiTransportError::Spi)?;
-
-        // We have now advertised the reset (if any) exactly once.
-        self.send_reset = false;
-
-        // ---- Parse the RCP's header. ----
-        let flag = self.rx_buf[0];
-        if flag == 0x00 || flag == 0xff {
-            // Bus idle / RCP not driving MISO — nothing exchanged this transfer.
-            return Ok(false);
-        }
-        if flag & SPI_FLAG_PATTERN_MASK != SPI_FLAG_PATTERN {
-            return Err(SpiTransportError::BadHeader);
-        }
-
-        let slave_accept_len = u16::from_le_bytes([self.rx_buf[1], self.rx_buf[2]]) as usize;
-        let slave_data_len = u16::from_le_bytes([self.rx_buf[3], self.rx_buf[4]]) as usize;
-
-        if slave_accept_len > MAX_SPINEL_FRAME || slave_data_len > MAX_SPINEL_FRAME {
-            return Err(SpiTransportError::BadHeader);
-        }
-
-        // Remember the RCP's pending length for sizing the *next* transfer.
-        self.slave_data_len = slave_data_len;
-
-        // ---- Did our outbound frame get accepted? ----
-        if tx_payload_len != 0 && tx_payload_len <= slave_accept_len {
-            self.tx_pending = None;
-        }
-
-        // ---- Did we receive a frame (and did it fit in our accept_len)? ----
-        let mut got_rx = false;
-        if slave_data_len != 0 && slave_data_len <= accept_len {
-            self.rx_frame[..slave_data_len]
-                .copy_from_slice(&self.rx_buf[SPI_HEADER_SIZE..SPI_HEADER_SIZE + slave_data_len]);
-            self.rx_ready = Some(slave_data_len);
-            // Consumed — the RCP will re-advertise if it has more.
-            self.slave_data_len = 0;
-            got_rx = true;
-        }
-
-        Ok(got_rx)
-    }
-}
-
-impl<S, I> SpinelTransport for SpiSpinelTransport<S, I>
-where
-    S: embedded_hal_async::spi::SpiDevice,
-    I: embedded_hal_async::digital::Wait,
-{
-    type Error = SpiTransportError<S::Error, I::Error>;
-
-    async fn send(&mut self, frame: &[u8]) -> Result<(), Self::Error> {
-        if frame.len() > MAX_SPINEL_FRAME {
-            return Err(SpiTransportError::FrameTooLarge);
-        }
-
-        // Stage the payload behind the header region and mark it pending.
-        self.tx_buf[SPI_HEADER_SIZE..SPI_HEADER_SIZE + frame.len()].copy_from_slice(frame);
-        self.tx_pending = Some(frame.len());
-
-        // Push transfers until the RCP accepts the frame. A transfer may also
-        // surface an inbound frame; buffer it for the next `recv`.
-        while self.tx_pending.is_some() {
-            self.push_pull().await?;
-        }
-
-        Ok(())
-    }
-
-    async fn recv(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        loop {
-            // Hand back an already-buffered frame first.
-            if let Some(len) = self.rx_ready.take() {
-                let len = len.min(buf.len());
-                buf[..len].copy_from_slice(&self.rx_frame[..len]);
-                return Ok(len);
-            }
-
-            // If the RCP has not yet told us it has data, wait for its interrupt.
-            // (If a prior header already announced a pending frame, go straight to
-            // a transfer to fetch it.)
-            if self.slave_data_len == 0 {
-                self.int
-                    .wait_for_low()
-                    .await
-                    .map_err(SpiTransportError::Int)?;
-            }
-
-            self.push_pull().await?;
-        }
-    }
-}
-
-/// Error type for [`SpiSpinelTransport`].
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum SpiTransportError<SpiErr, IntErr> {
-    /// The underlying SPI device returned an error.
-    Spi(SpiErr),
-    /// Waiting on the interrupt line failed.
-    Int(IntErr),
-    /// The RCP's SPI header failed the pattern/length sanity checks.
-    BadHeader,
-    /// A spinel frame exceeded the SPI transfer buffer.
-    FrameTooLarge,
 }
